# openapi/QAlbum.yaml
openapi: 3.0.3
info:
  title: QAlbum API
  description: 家庭私有云相册后端 API（家庭空间版）
  version: 1.0.0
  contact:
    name: QAlbum Team
servers:
  - url: https://album.example.com/api
    description: Production server via Cloudflare Tunnel
paths:
...
  /spaces:
    get:
      summary: 获取家庭空间列表
      description: 当前用户加入的所有家庭空间
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Space'

    post:
      summary: 创建家庭空间
      description: 创建者自动成为管理员
      tags: [Space]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, quota_bytes]
              properties:
                name:
                  type: string
                  maxLength: 50
                quota_bytes:
                  type: integer
                  description: 容量上限（字节）
                backup_path:
                  type: string
                  description: 可选，备份目标路径
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
  /spaces/{space_id}:
    get:
      summary: 获取家庭空间详情
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

    put:
      summary: 修改家庭空间（重命名、改配额、改备份路径）
      description: 仅管理员可操作
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                quota_bytes:
                  type: integer
                backup_path:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

    delete:
      summary: 删除家庭空间
      description: 仅管理员可删除，物理文件全部删除
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功
  /spaces/{space_id}/invite:
    post:
      summary: 生成邀请码
      description: 管理员生成邀请码，供家庭成员加入空间
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                expire_hours:
                  type: integer
                  default: 24
                  description: 邀请码有效期（小时）
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite_code:
                    type: string
                    description: 小程序卡片分享用的邀请码
                  expires_at:
                    type: string
                    format: date-time
  /spaces/join:
    post:
      summary: 通过邀请码加入家庭空间
      tags: [Space]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [invite_code]
              properties:
                invite_code:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'
  /spaces/{space_id}/members:
    get:
      summary: 获取家庭空间成员列表
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SpaceMember'
  /spaces/{space_id}/members/{user_id}:
    delete:
      summary: 移除成员
      description: 仅管理员可操作，不能移除自己
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
        - name: user_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 移除成功
  /spaces/{space_id}/backup/sync:
    post:
      summary: 手动触发备份同步（补备）
      description: 管理员触发，将空间内待同步文件复制到外接存储
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 同步已触发（异步执行）
          content:
            application/json:
              schema:
                type: object
                properties:
                  pending_count:
                    type: integer
                    description: 待同步文件数量

  # ─────────────────────────────── 相册 ───────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Space:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        owner_id:
          type: integer
          description: 创建者/管理员 ID
        quota_bytes:
          type: integer
        backup_path:
          type: string
          nullable: true
        backup_status:
          type: string
          enum: [ok, pending, unavailable]
        pending_backup_count:
          type: integer
          description: 待同步文件数量
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    SpaceMember:
      type: object
      description: 成员列表仅返回用户 ID 和角色，不暴露昵称/头像
      properties:
        user_id:
          type: integer
        role:
          type: string
          enum: [admin, member]
        joined_at:
          type: string
          format: date-time
...